make vinegar work in wide mode b:netrw_liststyle=2
In wide mode a file name does not necessarily begin on the beginning of a line neither  end on the end of a line.
Instead check for leading/trailing spaces.

Has the side effect that, when searching for a file A, if there is in the same directory another file B whose name extends that of A, then the search stops at B instead of A. There is no way to work around that in wide mode and therefore the pattern is only applied in that g:netrw_liststyle.