Add mkdir in Installation instructions
If this is the first bundle the user is installing, the directory ~/.vim/bundle may well not already exist, in which case the provided instructions would not at all achieve the desired result.  In fact the `cd` would fail and then `git clone` would drop a copy of pathogen in whatever directory the user happened to be in.

It might be better still to make the commands conditional on each other, as in `mkdir -p ~/.vim/bundle && cd ~/.vim/bundle && git clone ...` but for this PR I'm keeping it simple.