Remove prefix of tree view items on command
In order for the . , ! or y commands to work the intended way, the
leading '| ' will be now stripped when executing one of the above
commands.

There is the known limitation that this will not work with folders/files
starting with '| ' as this was considered an edge case and too
complicated for this easy fix.

Fix for #85